+++ xenon-common/src/main/java/com/vmware/xenon/common/UtilityService.java
@@ -31,6 +31,8 @@ import com.vmware.xenon.common.Operation.CompletionHandler;
 import com.vmware.xenon.common.ServiceStats.ServiceStat;
 import com.vmware.xenon.common.ServiceSubscriptionState.ServiceSubscriber;
 import com.vmware.xenon.services.common.ServiceUriPaths;
+import com.vmware.xenon.services.common.UiContentService;
+
 
 /**
  * Utility service managing the various URI control REST APIs for each service instance. A single
@@ -41,6 +43,7 @@ public class UtilityService implements Service {
     private transient Service parent;
     private ServiceStats stats;
     private ServiceSubscriptionState subscriptions;
+    private UiContentService uiService;
 
     public UtilityService() {
     }
@@ -327,89 +330,34 @@ public class UtilityService implements Service {
             String defaultHtmlPath = UriUtils.buildUriPath(servicePath.substring(0,
                     servicePath.length() - ServiceUriPaths.UI_PATH_SUFFIX.length()),
                     ServiceUriPaths.UI_SERVICE_HOME);
-            try {
-                redirectGetToHtmlUiResource(op, defaultHtmlPath);
-            } catch (UnsupportedEncodingException e) {
-                op.fail(e);
-            }
+
+            redirectGetToHtmlUiResource(op, defaultHtmlPath);
             return;
         }
 
-        // The entry point to the UI resource rendering is a HTML file named the same as the service
-        // class.
-        // E.g com.vmware.ExampleService -> com/vmware/ExampleService/index.html
-        String serviceUiResourcePath;
-        if (this.parent.getDocumentTemplate().documentDescription != null &&
-                this.parent.getDocumentTemplate().documentDescription.userInterfaceResourcePath
-                    != null) {
-            serviceUiResourcePath = Utils.buildCustomUiResourceUriPrefixPath(this.parent);
-        } else {
-            serviceUiResourcePath = Utils.buildUiResourceUriPrefixPath(this.parent);
+        if (this.uiService == null) {
+            this.uiService = new UiContentService() {
+            };
+            this.uiService.setHost(this.parent.getHost());
         }
 
-        // find what's after the selfLink
-        String uriPath = op.getUri().getPath();
-        String selfLink = this.parent.getSelfLink();
-        uriPath = uriPath.substring(selfLink.length() + ServiceHost.SERVICE_URI_SUFFIX_UI.length());
-
-        if (uriPath.equals("")) {
-            try {
-                // redirect to /service/ui/ (trailing slash!)
-                redirectGetToTrailingSlash(op, selfLink + ServiceHost.SERVICE_URI_SUFFIX_UI);
-                return;
-            } catch (UnsupportedEncodingException e) {
-                op.fail(e);
-                return;
-            }
-        } else if (uriPath.equals(UriUtils.URI_PATH_CHAR)) {
-            // serve index.html on /service/ui/
-            serviceUiResourcePath += UriUtils.URI_PATH_CHAR + ServiceUriPaths.UI_RESOURCE_DEFAULT_FILE;
-        } else {
-            // forward to /service/ui/some-resource.js
-            serviceUiResourcePath += uriPath;
-        }
-
-        proxyGetToCustomHtmlUiResource(op, serviceUiResourcePath);
+        // simulate a full service deployed at the utility endpoint /service/ui
+        String selfLink = this.parent.getSelfLink() + ServiceHost.SERVICE_URI_SUFFIX_UI;
+        this.uiService.handleUiGet(selfLink, this.parent, op);
     }
 
-    public void redirectGetToHtmlUiResource(Operation op, String htmlResourcePath)
-            throws UnsupportedEncodingException {
+    public void redirectGetToHtmlUiResource(Operation op, String htmlResourcePath) {
         // redirect using relative url without host:port
         // not so much optimization as handling the case of port forwarding/containers
-        op.addResponseHeader(Operation.LOCATION_HEADER,
-                URLDecoder.decode(htmlResourcePath, Utils.CHARSET));
+        try {
+            op.addResponseHeader(Operation.LOCATION_HEADER,
+                    URLDecoder.decode(htmlResourcePath, Utils.CHARSET));
+        } catch (UnsupportedEncodingException e) {
+            throw new IllegalStateException(e);
+        }
+
         op.setStatusCode(Operation.STATUS_CODE_MOVED_TEMP);
-        op.setContentType(Operation.MEDIA_TYPE_TEXT_HTML);
         op.complete();
-    }
-
-    /**
-     * 302 redirect to the provided folderPath with a '/' appended
-     * @param op
-     * @param folderPath
-     * @throws UnsupportedEncodingException
-     */
-    private void redirectGetToTrailingSlash(Operation op, String folderPath)
-            throws UnsupportedEncodingException {
-        op.addResponseHeader(Operation.LOCATION_HEADER,
-                URLDecoder.decode(folderPath + UriUtils.URI_PATH_CHAR, Utils.CHARSET));
-        op.setStatusCode(Operation.STATUS_CODE_MOVED_TEMP);
-        op.setContentType(Operation.MEDIA_TYPE_TEXT_HTML);
-        op.complete();
-    }
-
-    public void proxyGetToCustomHtmlUiResource(Operation op, String htmlResourcePath) {
-        Operation get = op.clone();
-        get.setUri(UriUtils.buildUri(getHost(), htmlResourcePath))
-                .setReferer(op.getReferer())
-                .setCompletion((o, e) -> {
-                    op.setBody(o.getBodyRaw())
-                            .setContentType(o.getContentType())
-                            .complete();
-                });
-
-        getHost().sendRequest(get);
-        return;
     }
 
     private void handleStatsRequest(Operation op) {