+++ xenon-samples/src/test/java/com/vmware/dcp/services/samples/TestSampleServiceWithCustomUi.java
@@ -27,6 +27,7 @@ import com.vmware.xenon.common.Operation;
 import com.vmware.xenon.common.UriUtils;
 import com.vmware.xenon.services.samples.SampleFactoryServiceWithCustomUi;
 import com.vmware.xenon.services.samples.SampleServiceWithCustomUi.SampleServiceWithCustomUiState;
+import com.vmware.xenon.services.samples.SampleServiceWithSharedCustomUi;
 
 public class TestSampleServiceWithCustomUi extends BasicReusableHostTestCase {
 
@@ -34,10 +35,10 @@ public class TestSampleServiceWithCustomUi extends BasicReusableHostTestCase {
 
     @Before
     public void prepare() throws Throwable {
-        URI factoryUri = UriUtils.buildUri(this.host, SampleFactoryServiceWithCustomUi.class);
-        this.host.startService(
-                Operation.createPost(factoryUri),
-                new SampleFactoryServiceWithCustomUi());
+        this.host.startService(new SampleFactoryServiceWithCustomUi());
+
+        // start the service that serves the ui resources
+        this.host.startService(new SampleServiceWithSharedCustomUi());
 
         createService();
     }
@@ -77,6 +78,36 @@ public class TestSampleServiceWithCustomUi extends BasicReusableHostTestCase {
     }
 
     @Test
+    public void testRootGetForSharedUi() throws Throwable {
+        Operation op = Operation
+                .createGet(UriUtils.buildUri(this.host, SampleServiceWithSharedCustomUi.SELF_LINK))
+                .setCompletion(getSafeHandler((o, e) -> {
+                    assertNull(e);
+                    assertEquals("Did not receive temporary redirect",  Operation.STATUS_CODE_MOVED_TEMP, o.getStatusCode());
+                    assertTrue("Redirected url does not end with /", o.getResponseHeader(Operation.LOCATION_HEADER).endsWith("/"));
+                }));
+
+        this.host.testStart(1);
+        this.host.send(op);
+        this.host.testWait();
+    }
+
+    @Test
+    public void testGetUiWithSlashForSharedUi() throws Throwable {
+        Operation op = Operation
+                .createGet(URI.create(UriUtils.buildUri(this.host, SampleServiceWithSharedCustomUi.SELF_LINK) + "/"))
+                .setCompletion(getSafeHandler((o, e) -> {
+                    assertNull(e);
+                    assertEquals("Expected 200 OK", Operation.STATUS_CODE_OK, o.getStatusCode());
+                    assertTrue("Expected content of the index.html", o.getBody(String.class).contains("customUiApp"));
+                }));
+
+        this.host.testStart(1);
+        this.host.send(op);
+        this.host.testWait();
+    }
+
+    @Test
     public void testGetUiWithSlash() throws Throwable {
         Operation op = Operation
                 .createGet(URI.create(UriUtils.buildUri(this.host, THE_SERVICE_URI + "/ui") + "/"))
@@ -92,12 +123,28 @@ public class TestSampleServiceWithCustomUi extends BasicReusableHostTestCase {
     }
 
     @Test
+    public void testRelativePathGetForSharedUi() throws Throwable {
+        Operation op = Operation
+                .createGet(UriUtils.buildUri(this.host,
+                        SampleServiceWithSharedCustomUi.SELF_LINK + "/constants.js"))
+                .setCompletion(getSafeHandler((o, e) -> {
+                    assertNull(e);
+                    assertEquals("Expected 200 OK", Operation.STATUS_CODE_OK, o.getStatusCode());
+                    assertTrue("Expected the contents of the README.txt", o.getBody(String.class).contains("UI_CUSTOM_BASE"));
+                }));
+
+        this.host.testStart(1);
+        this.host.send(op);
+        this.host.testWait();
+    }
+
+    @Test
     public void testGetResource() throws Throwable {
         Operation op = Operation
                 .createGet(URI.create(UriUtils.buildUri(this.host, THE_SERVICE_URI + "/ui/README.txt") + "/"))
                 .setCompletion(getSafeHandler((o, e) -> {
                     assertNull(e);
-                    assertEquals("Expected 200 OK",Operation.STATUS_CODE_OK, o.getStatusCode());
+                    assertEquals("Expected 200 OK", Operation.STATUS_CODE_OK, o.getStatusCode());
                     assertTrue("Expected the contents of the README.txt", o.getBody(String.class).contains("unit-test"));
                 }));